% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_n_steps}
\alias{compute_n_steps}
\title{Compute The Number of Model Steps}
\usage{
compute_n_steps(n_times, step_interval)
}
\arguments{
\item{n_times}{\code{\link[=integer]{integer()}} number of model times (years)}

\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\code{\link[=integer]{integer()}} number of model steps
}
\description{
Compute The Number of Model Steps
}
\examples{
compute_n_steps(1, "month")
compute_n_steps(1, "quarter")
compute_n_steps(1, "year")
compute_n_steps(10, "month")
compute_n_steps(10, "quarter")
compute_n_steps(10, "year")

}
