% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_n_terms}
\alias{compute_n_terms}
\title{Compute The Number of Terms Per Year}
\usage{
compute_n_terms(x)
}
\arguments{
\item{x}{\code{\link[=character]{character()}} one of \code{"year"}, \code{"quarter"}
or \code{"month"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Compute The Number of Terms Per Year
}
\examples{

compute_n_terms("year")
compute_n_terms("quarter")
compute_n_terms("month")

}
