% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_n_times}
\alias{compute_n_times}
\title{Compute The Number of Years}
\usage{
compute_n_times(a, b)
}
\arguments{
\item{a}{\code{\link[=character]{character()}} date as \code{"\%Y-\%m-\%d"}}

\item{b}{\code{\link[=character]{character()}} date as \code{"\%Y-\%m-\%d"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Compute The Number of Years
}
\examples{

compute_n_times("2011-01-01", "2014-12-31")

}
