% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_steps_per_term}
\alias{compute_steps_per_term}
\title{Compute Steps Per Term}
\usage{
compute_steps_per_term(step_interval, term_interval)
}
\arguments{
\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}

\item{term_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Compute Steps Per Term
}
\examples{
compute_steps_per_term("month", "quarter")
compute_steps_per_term("quarter", "quarter")
compute_steps_per_term("quarter", "year")

}
