% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_terms_per_year}
\alias{compute_terms_per_year}
\title{Compute The Number of Model Terms Per Year}
\usage{
compute_terms_per_year(x)
}
\arguments{
\item{x}{\code{\link[=character]{character()}} one of \code{"year"}, \code{"quarter"}
or \code{"month"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Compute The Number of Model Terms Per Year
}
\examples{

compute_terms_per_year("year")
compute_terms_per_year("quarter")
compute_terms_per_year("month")

}
