% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_movement_steps}
\alias{convert_movement_steps}
\title{Convert Movement Steps}
\usage{
convert_movement_steps(model_form, step_interval)
}
\arguments{
\item{model_form}{\code{\link[=character]{character()}} one of \code{"mean", "time", "term", "size"}}

\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Convert Movement Steps
}
\examples{

convert_movement_steps("mean", "month")
convert_movement_steps("mean", "quarter")
convert_movement_steps("mean", "year")
convert_movement_steps("term", "month")
convert_movement_steps("term", "quarter")

}
