% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_stepwise_rates}
\alias{convert_stepwise_rates}
\title{Convert Stepwise Rates}
\usage{
convert_stepwise_rates(step_interval)
}
\arguments{
\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Convert Stepwise Rates
}
\examples{

convert_stepwise_rates("month")
convert_stepwise_rates("quarter")
convert_stepwise_rates("year")

}
