% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_intervals}
\alias{count_intervals}
\title{Count Intervals}
\usage{
count_intervals(a, b)
}
\arguments{
\item{a}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}}

\item{b}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Count Intervals
}
\examples{

count_intervals("month", "month")
count_intervals("month", "quarter")
count_intervals("month", "year")
count_intervals("quarter", "quarter")
count_intervals("quarter", "year")
count_intervals("year", "year")

}
