% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_model_steps}
\alias{count_model_steps}
\title{Count Model Steps}
\usage{
count_model_steps(year_start, year_end, step_interval)
}
\arguments{
\item{year_start}{\code{\link[=integer]{integer()}} year of first tag release}

\item{year_end}{\code{\link[=integer]{integer()}} year of final tag recovery}

\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Count Model Steps
}
\examples{

count_model_steps(2011, 2018, "month")
count_model_steps(2011, 2018, "quarter")
count_model_steps(2011, 2018, "year")

}
