% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_movement_sizes}
\alias{count_movement_sizes}
\title{Count Movement Sizes}
\usage{
count_movement_sizes(model_form, list_sizes)
}
\arguments{
\item{model_form}{\code{\link[=character]{character()}} one of \code{"mean", "time", "term", "size"}}

\item{list_sizes}{\code{\link[=list]{list()}}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Count Movement Sizes
}
\examples{

count_movement_sizes("mean", list(s = 1:5, m = 6:10, l = 11:15))
count_movement_sizes("time", list(s = 1:5, m = 6:10, l = 11:15))
count_movement_sizes("term", list(s = 1:5, m = 6:10, l = 11:15))
count_movement_sizes("size", list(s = 1:5, m = 6:10, l = 11:15))

}
