% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_movement_steps}
\alias{count_movement_steps}
\title{Count Movement Steps}
\usage{
count_movement_steps(model_form, year_start, year_end)
}
\arguments{
\item{model_form}{\code{\link[=character]{character()}} one of \code{"mean", "time", "term", "size"}}

\item{year_start}{\code{\link[=integer]{integer()}} year of first tag release}

\item{year_end}{\code{\link[=integer]{integer()}} year of final tag recovery}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Count Movement Steps
}
\examples{

count_movement_steps("mean", 2011, 2018)
count_movement_steps("time", 2011, 2018)
count_movement_steps("term", 2011, 2018)

}
