% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_steps_per_year}
\alias{count_steps_per_year}
\title{Count Model Steps Per Year}
\usage{
count_steps_per_year(step_interval)
}
\arguments{
\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Count Model Steps Per Year
}
\examples{

count_steps_per_year("month")
count_steps_per_year("quarter")
count_steps_per_year("year")

}
