% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_beta_parameters}
\alias{create_beta_parameters}
\title{Create Beta Parameters}
\usage{
create_beta_parameters(mu, sd, n = 10000, b = 100)
}
\arguments{
\item{mu}{\code{\link[=numeric]{numeric()}} Mean}

\item{sd}{\code{\link[=numeric]{numeric()}} Standard deviation}

\item{n}{\code{\link[=integer]{integer()}} Number of samples for histogram}

\item{b}{\code{\link[=numeric]{numeric()}} Breaks passed to \code{\link[=hist]{hist()}}}
}
\value{
\code{\link[=numeric]{numeric()}} vector of length two.
}
\description{
Create Beta Parameters
}
\examples{
create_beta_parameters(0.01, 0.001)

}
