% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_gamma_parameters}
\alias{create_gamma_parameters}
\title{Create Gamma Parameters}
\usage{
create_gamma_parameters(mu, sd)
}
\arguments{
\item{mu}{\code{\link[=numeric]{numeric()}} \code{\link[=array]{array()}} giving the mean(s).}

\item{sd}{\code{\link[=numeric]{numeric()}} shared standard deviation.}
}
\value{
\code{\link[=list]{list()}} of two arrays
}
\description{
Create Gamma Parameters
}
\examples{
# Scalar mu
mu <- 0.05
l <- create_gamma_parameters(mu, 0.01)
hist(rgamma(10000, l$alpha, l$beta), breaks = 100)

# Array mu
mu <- array(0.05, dim = c(2, 10, 3))
l <- create_gamma_parameters(mu, 0.001)
hist(rgamma(10000, l$alpha[1,1,1], l$beta[1,1,1]), breaks = 100)

}
