% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_group}
\alias{create_group}
\title{Create Groups from Index List}
\usage{
create_group(x, list_x)
}
\arguments{
\item{x}{\code{\link[=atomic]{atomic()}} \code{\link[=vector]{vector()}} of values}

\item{list_x}{\code{\link[=list]{list()}} of groups of x}
}
\value{
\code{\link[=integer]{integer()}} \code{\link[=vector]{vector()}} of numeric index of groups
}
\description{
Create Groups from Index List
}
\examples{

# Numeric
x <- c(1, 4, 2, 3, NA, 2, 7)
g <- list(a = 1:2, b = 3:4)
create_group(x, g)

# Character
x <- c("M", "F", "F", NA, "M", "N", "F")
g <- list(m = "M", f = "F")
create_group(x, g)

# Factor
x <- factor(c("L", "S", "M", NA, "M", "H"))
g <- list(s = "S", m = "M", l = "L")
create_group(x, g)

}
