% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_h_step}
\alias{create_h_step}
\title{Create Harvest Rate Step}
\usage{
create_h_step(h_annual, y)
}
\arguments{
\item{h_annual}{\code{\link[=numeric]{numeric()}} annual harvest rate}

\item{y}{\code{\link[=integer]{integer()}} number of time steps in one year}
}
\value{
\code{\link[=numeric]{numeric()}}
}
\description{
Create Harvest Rate Step
}
\examples{
create_h_step(0.05, 1)
create_h_step(0.05, 4)

}
