% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_movement_index}
\alias{create_movement_index}
\title{Create A Movement Index Matrix}
\usage{
create_movement_index(
  n_regions,
  movement_pattern = 2,
  movement_allow = NULL,
  movement_disallow = NULL
)
}
\arguments{
\item{n_regions}{\code{\link[=integer]{integer()}} Number of regions}

\item{movement_pattern}{\code{\link[=integer]{integer()}} One of \code{1}: movement between
all pairs of regions at each step, or \code{2}:
movement between numerically sequential regions (the default)}

\item{movement_allow}{\code{\link[=integer]{integer()}} \code{\link[=matrix]{matrix()}} Adjusts the movement pattern
to allow additional movement between regions. Each row indicates
directional movement between a pair of regions}

\item{movement_disallow}{\code{\link[=integer]{integer()}} \code{\link[=matrix]{matrix()}} As for
\code{movement_allow}, but specified movement is disallowed}
}
\value{
A square matrix of zeros and ones
}
\description{
Create A Movement Index Matrix
}
\examples{

# Neighbors (default)
create_movement_index(6)

# All pairwise movement
create_movement_index(6, 1)

# Neighbors plus 1-6, 2-5, and 5-2, but not 6-1
movement_allow <- matrix(c(1,6,2,5,5,2), ncol = 2, byrow = TRUE)
create_movement_index(6, 2, movement_allow)

# All pairwise movement but not 1-6
movement_disallow <- matrix(c(1,6), ncol = 2, byrow = TRUE)
create_movement_index(6, 1, movement_disallow = movement_disallow)

}
