% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_n_to_i}
\alias{create_n_to_i}
\title{Create N to I Index}
\usage{
create_n_to_i(model_form, year_start, year_end, step_interval)
}
\arguments{
\item{model_form}{\code{\link[=character]{character()}} one of \code{"mean", "time", "term", "size"}}

\item{year_start}{\code{\link[=integer]{integer()}} year of first tag release}

\item{year_end}{\code{\link[=integer]{integer()}} year of final tag recovery}

\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Create N to I Index
}
\examples{

create_n_to_i("mean", 2011, 2018, "month")
create_n_to_i("mean", 2011, 2018, "quarter")
create_n_to_i("mean", 2011, 2018, "year")
create_n_to_i("time", 2011, 2018, "month")
create_n_to_i("time", 2011, 2018, "quarter")
create_n_to_i("time", 2011, 2018, "year")
create_n_to_i("term", 2011, 2018, "month")
create_n_to_i("term", 2011, 2018, "quarter")
create_n_to_i("size", 2011, 2018, "month")
create_n_to_i("size", 2011, 2018, "quarter")
create_n_to_i("size", 2011, 2018, "year")

}
