% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_n_to_t}
\alias{create_n_to_t}
\title{Create N to T Index}
\usage{
create_n_to_t(year_start, year_end, step_interval)
}
\arguments{
\item{year_start}{\code{\link[=integer]{integer()}} year of first tag release}

\item{year_end}{\code{\link[=integer]{integer()}} year of final tag recovery}

\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Create N to T Index
}
\examples{

create_n_to_t(2011, 2018, "month")
create_n_to_t(2011, 2018, "quarter")
create_n_to_t(2011, 2018, "year")

}
