% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_s_to_d}
\alias{create_s_to_d}
\title{Create S to D Index Array}
\usage{
create_s_to_d(model_form, list_sizes)
}
\arguments{
\item{model_form}{\code{\link[=character]{character()}} one of \code{"mean", "time", "term", "size"}}

\item{list_sizes}{\code{\link[=list]{list()}}}
}
\value{
\code{\link[=integer]{integer()}} \code{\link[=vector]{vector()}}
}
\description{
Create S to D Index Array
}
\examples{

create_s_to_d("mean", list(s = 1:5, m = 6:10, l = 11:15))
create_s_to_d("time", list(s = 1:5, m = 6:10, l = 11:15))
create_s_to_d("term", list(s = 1:5, m = 6:10, l = 11:15))
create_s_to_d("size", list(s = 1:5, m = 6:10, l = 11:15))

}
