% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_sample_array}
\alias{create_sample_array}
\title{Create MCMC Sample Array}
\usage{
create_sample_array(fit, par_name)
}
\arguments{
\item{fit}{\code{\link[=mmmfit]{mmmfit()}} object}

\item{par_name}{\code{\link[=character]{character()}} parameter name (\code{p}, \code{h}, or
\code{phi})}
}
\value{
\code{\link[=array]{array()}}
}
\description{
Create MCMC Sample Array
}
