% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_sample_summary}
\alias{create_sample_summary}
\title{Create MCMC Sample Summary}
\usage{
create_sample_summary(x, ci_level = 0.8, outer_level = 0.95)
}
\arguments{
\item{x}{\code{\link[=array]{array()}} sample array (see \code{create_sample_array()})}

\item{ci_level}{\code{\link[=numeric]{numeric()}} scalar in c(0, 1)}

\item{outer_level}{\code{\link[=numeric]{numeric()}} scalar in c(0, 1)}
}
\value{
\code{\link[=array]{array()}}
}
\description{
Create MCMC Sample Summary
}
