% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_sample_tibble}
\alias{create_sample_tibble}
\title{Create MCMC Sample Tibble}
\usage{
create_sample_tibble(fit, pars = c("p", "h", "phi"))
}
\arguments{
\item{fit}{\code{\link[=mmmfit]{mmmfit()}} object}

\item{pars}{\code{\link[=character]{character()}} \code{\link[=vector]{vector()}} parameter names (\code{p}, \code{h},
and/or \code{phi})}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create MCMC Sample Tibble
}
