% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_step_liberty}
\alias{create_step_liberty}
\title{Create Step Liberty}
\usage{
create_step_liberty(
  date_released,
  date_recovered,
  date_released_start,
  date_recovered_end,
  term_interval,
  step_liberty_max
)
}
\arguments{
\item{date_released}{\code{\link[=character]{character()}} \code{\link[=vector]{vector()}} dates as \code{"\%Y-\%m-\%d"}}

\item{date_recovered}{\code{\link[=character]{character()}} \code{\link[=vector]{vector()}} dates as \code{"\%Y-\%m-\%d"}}

\item{date_released_start}{\code{\link[=character]{character()}} date as \code{"\%Y-\%m-\%d"}}

\item{date_recovered_end}{\code{\link[=character]{character()}} date as \code{"\%Y-\%m-\%d"}}

\item{term_interval}{\code{\link[=character]{character()}} one of \code{"year"}, \code{"quarter"}
or \code{"month"}}

\item{step_liberty_max}{\code{\link[=integer]{integer()}} maximum steps at liberty}
}
\value{
\code{\link[=integer]{integer()}} \code{\link[=vector]{vector()}} step liberty
}
\description{
Create Step Liberty
}
\examples{

# Yearly
u <- rep("2011-01-01", 5)
v <- c("2011-01-01", "2012-01-01", NA, "2020-01-01", "2021-01-01")
w <- c("2011-01-01")
x <- c("2020-01-01")
y  <- "year"
z <- 10
create_step_liberty(u, v, w, x, y, z)

# Quarterly
u <- rep("2011-01-01", 5)
v <- c("2011-01-01", "2012-12-31", NA, "2020-12-31", "2021-01-01")
w <- c("2011-01-01")
x <- c("2020-01-01")
y <- "quarter"
z <- 40
create_step_liberty(u, v, w, x, y, z)

# Monthly
u <- rep("2011-01-01", 5)
v <- c("2011-01-01", "2011-02-01", NA, "2020-12-31", "2021-01-01")
w <- c("2011-01-01")
x <- c("2020-01-01")
y <- "month"
z <- 120
create_step_liberty(u, v, w, x, y, z)

}
