% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_step_released}
\alias{create_step_released}
\title{Create Step Released}
\usage{
create_step_released(
  date_released,
  date_released_start,
  date_recovered_end,
  term_interval
)
}
\arguments{
\item{date_released}{\code{\link[=character]{character()}} \code{\link[=vector]{vector()}} dates as \code{"\%Y-\%m-\%d"}}

\item{date_released_start}{\code{\link[=character]{character()}} date as \code{"\%Y-\%m-\%d"}}

\item{date_recovered_end}{\code{\link[=character]{character()}} date as \code{"\%Y-\%m-\%d"}}

\item{term_interval}{\code{\link[=character]{character()}} one of \code{"year"}, \code{"quarter"}
or \code{"month"}}
}
\value{
\code{\link[=integer]{integer()}} \code{\link[=vector]{vector()}} step released
}
\description{
Create Step Released
}
\examples{

x <- c("2010-01-01", "2011-01-01", NA, "2020-12-31", "2021-01-01")
d0 <- c("2011-01-01")
d1 <- c("2020-12-31")
create_step_released(x, d0, d1, "year")
create_step_released(x, d0, d1, "quarter")
create_step_released(x, d0, d1, "month")

}
