% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_step_to_rate_power}
\alias{create_step_to_rate_power}
\title{Create Step to Rate Power}
\usage{
create_step_to_rate_power(step_interval, rate_interval)
}
\arguments{
\item{step_interval}{\code{\link[=character]{character()}} in \code{c("month", "quarter", "year")}}

\item{rate_interval}{\code{\link[=character]{character()}} in \code{c("month", "quarter", "year")}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Create Step to Rate Power
}
\examples{

create_step_to_rate_power("month", "month")
create_step_to_rate_power("month", "quarter")
create_step_to_rate_power("month", "year")
create_step_to_rate_power("quarter", "quarter")
create_step_to_rate_power("quarter", "year")
create_step_to_rate_power("year", "year")

}
