% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_step_to_term}
\alias{create_step_to_term}
\title{Create Step to Term Index}
\usage{
create_step_to_term(n_steps, n_times, n_terms, nest_terms_within_times = FALSE)
}
\arguments{
\item{n_steps}{\code{\link[=integer]{integer()}} number of model steps}

\item{n_times}{\code{\link[=integer]{integer()}} number of times (years)}

\item{n_terms}{\code{\link[=integer]{integer()}} number of unique terms (seasons)}

\item{nest_terms_within_times}{\code{\link[=logical]{logical()}} nest terms within times?}
}
\value{
\link[=vector]{integer()}
}
\description{
Create Step to Term Index
}
