% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_step_to_time}
\alias{create_step_to_time}
\title{Create Step to Time Index}
\usage{
create_step_to_time(n_steps, n_times)
}
\arguments{
\item{n_steps}{\code{\link[=integer]{integer()}} number of model steps}

\item{n_times}{\code{\link[=integer]{integer()}} number of times (years)}
}
\value{
\link[=vector]{integer()}
}
\description{
Create Step to Time Index
}
