% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_tag_areas}
\alias{create_tag_areas}
\title{Create Tag Areas}
\usage{
create_tag_areas(x, area_list)
}
\arguments{
\item{x}{\code{\link[=atomic]{atomic()}} vector of area identifiers}

\item{area_list}{\code{\link[=list]{list()}} of named vectors of area identifiers}
}
\value{
\code{\link[=integer]{integer()}} vector
}
\description{
Create Tag Areas
}
\examples{

# Numeric area identifiers
x <- c(1, 4, 2, 3, NA, 2, 7)
a <- list(area_1 = 1:2, area_2 = 3:4)
create_tag_areas(x, a)

# Character area identifiers
x <- c("A", "B", "B", NA, "A", "C", "B")
a <- list(area_1 = "A", area_2 = "B")
create_tag_areas(x, a)

# Factor
x <- factor(c("C", "A", "B", NA, "B", "D"))
a <- list(area_1 = "A", area_2 = "B", area_3 = "C")
create_tag_areas(x, a)

}
