% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_tag_array}
\alias{create_tag_array}
\title{Create Tag Array}
\usage{
create_tag_array(
  x,
  y,
  time_step,
  max_steps_liberty,
  release_date_colname,
  release_area_colname,
  group_colname,
  recovery_date_colname,
  recovery_area_colname,
  id_colname,
  area_list,
  group_list,
  release_date_start,
  release_date_end
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}} tag release observation data. See details}

\item{y}{\code{\link[=data.frame]{data.frame()}} tag recovery observation data. See details}

\item{time_step}{\code{\link[=character]{character()}} desired time step for tag array. One of
\code{c("year", "month")}}

\item{max_steps_liberty}{\code{\link[=integer]{integer()}} maximum time steps at liberty}

\item{release_date_colname}{\code{\link[=character]{character()}} column name of the release date
in \code{x} and \code{y}}

\item{release_area_colname}{\code{\link[=character]{character()}} column name of the release area
in \code{x} and \code{y}}

\item{group_colname}{\code{\link[=character]{character()}} column name of the group indicator
in \code{x} and \code{y}}

\item{recovery_date_colname}{\code{\link[=character]{character()}} column name of the recovery date
in \code{y}}

\item{recovery_area_colname}{\code{\link[=character]{character()}} column name of the recovery area
in \code{y}}

\item{id_colname}{\code{\link[=character]{character()}} column name of the tag identifier
in \code{x} and \code{y}}

\item{area_list}{\code{\link[=list]{list()}} named atomic vectors. See details}

\item{group_list}{\code{\link[=list]{list()}} named atomic vectors. See details}

\item{release_date_start}{\code{\link[=character]{character()}} release start date}

\item{release_date_end}{\code{\link[=character]{character()}} release end date}
}
\value{
\code{\link[=array]{array()}} aggregated tag observation data.
}
\description{
Create Tag Array
}
\details{
TBD
}
