% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_tag_groups}
\alias{create_tag_groups}
\title{Create Tag Groups}
\usage{
create_tag_groups(x, group_list)
}
\arguments{
\item{x}{\code{\link[=atomic]{atomic()}} vector of values to be grouped}

\item{group_list}{\code{\link[=list]{list()}} of named vectors of group elements}
}
\value{
\code{\link[=integer]{integer()}} vector
}
\description{
Create Tag Groups
}
\examples{

# Numeric
x <- c(1, 4, 2, 3, NA, 2, 7)
g <- list(a = 1:2, b = 3:4)
create_tag_groups(x, g)

# Character
x <- c("M", "F", "F", NA, "M", "N", "F")
g <- list(m = "M", f = "F")
create_tag_groups(x, g)

# Factor
x <- factor(c("L", "S", "M", NA, "M", "H"))
g <- list(s = "S", m = "M", l = "L")
create_tag_groups(x, g)

}
