% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_tag_liberty_steps}
\alias{create_tag_liberty_steps}
\title{Create Tag Liberty Steps}
\usage{
create_tag_liberty_steps(
  release_date,
  recovery_date,
  release_start,
  release_end,
  time_step,
  max_steps_liberty
)
}
\arguments{
\item{release_date}{\code{\link[=character]{character()}} vector of release dates}

\item{recovery_date}{\code{\link[=character]{character()}} vector of recovery dates}

\item{release_start}{\code{\link[=character]{character()}} release date start}

\item{release_end}{\code{\link[=character]{character()}} release date end}

\item{time_step}{\code{\link[=character]{character()}} one of \code{c("year", "quarter", "month")}}

\item{max_steps_liberty}{\code{\link[=integer]{integer()}} maximum number of time steps at liberty}
}
\value{
\code{\link[=integer]{integer()}} vector
}
\description{
Create Tag Liberty Steps
}
\examples{

# Yearly
rel_date <- rep("2011-01-01", 5)
rec_date <- c("2011-01-01", "2012-01-01", NA, "2020-01-01", "2021-01-01")
rel_start <- c("2011-01-01")
rel_end <- c("2020-01-01")
time_step <- "year"
max_steps_liberty <- 10
create_tag_liberty_steps(rel_date,
                         rec_date,
                         rel_start,
                         rel_end,
                         time_step,
                         max_steps_liberty)

# Quarterly
rel_date <- rep("2011-01-01", 5)
rec_date <- c("2011-01-01", "2012-12-31", NA, "2020-12-31", "2021-01-01")
rel_start <- c("2011-01-01")
rel_end <- c("2020-01-01")
time_step <- "quarter"
max_steps_liberty <- 40
create_tag_liberty_steps(rel_date,
                         rec_date,
                         rel_start,
                         rel_end,
                         time_step,
                         max_steps_liberty)

# Monthly
rel_date <- rep("2011-01-01", 5)
rec_date <- c("2011-01-01", "2011-02-01", NA, "2020-12-31", "2021-01-01")
rel_start <- c("2011-01-01")
rel_end <- c("2020-01-01")
time_step <- "month"
max_steps_liberty <- 120
create_tag_liberty_steps(rel_date,
                         rec_date,
                         rel_start,
                         rel_end,
                         time_step,
                         max_steps_liberty)

}
