% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_tag_release_steps}
\alias{create_tag_release_steps}
\title{Create Tag Release Step}
\usage{
create_tag_release_steps(release_date, release_start, release_end, time_step)
}
\arguments{
\item{release_date}{\code{\link[=character]{character()}} vector of release dates}

\item{release_start}{\code{\link[=character]{character()}} release date start}

\item{release_end}{\code{\link[=character]{character()}} release date end}

\item{time_step}{\code{\link[=character]{character()}} one of \code{c("year", "quarter", "month")}}
}
\value{
\code{\link[=integer]{integer()}} vector of tag release time steps
}
\description{
Create Tag Release Step
}
\examples{

x <- c("2010-01-01", "2011-01-01", NA, "2020-12-31", "2021-01-01")
d0 <- c("2011-01-01")
d1 <- c("2020-12-31")
create_tag_release_steps(x, d0, d1, "year")
create_tag_release_steps(x, d0, d1, "quarter")
create_tag_release_steps(x, d0, d1, "month")

}
