% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_tags_released}
\alias{create_tags_released}
\title{Create Tags Released}
\usage{
create_tags_released(
  tags,
  list_regions,
  list_sizes,
  date_released_start,
  date_released_end,
  term_released = "quarter",
  colname_date_released = "date_released",
  colname_region_released = "region_released",
  colname_size_released = "size_released"
)
}
\arguments{
\item{tags}{\code{\link[=data.frame]{data.frame()}}}

\item{list_regions}{\code{\link[=list]{list()}}}

\item{list_sizes}{\code{\link[=list]{list()}}}

\item{date_released_start}{\code{\link[=character]{character()}} date as \code{"\%Y-\%m-\%d"}}

\item{date_released_end}{\code{\link[=character]{character()}} date as \code{"\%Y-\%m-\%d"}}

\item{term_released}{\code{\link[=character]{character()}} one of \code{"year"}, \code{"quarter"}
or \code{"month"}}

\item{colname_date_released}{\code{\link[=character]{character()}}}

\item{colname_region_released}{\code{\link[=character]{character()}}}

\item{colname_size_released}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create Tags Released
}
