% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit}
\alias{fit}
\title{Fit Markov Movement Model via CmdStanR}
\usage{
fit(
  data = NULL,
  chains = 1,
  step_size = 0.01,
  iter_warmup = 250,
  iter_sampling = 750,
  use_reduce_sum = FALSE,
  threads_per_chain = 8,
  ...
)
}
\arguments{
\item{data}{\code{\link[=list]{list()}} See details}

\item{chains}{\code{\link[=integer]{integer()}} number of chains}

\item{step_size}{\code{\link[=numeric]{numeric()}} initial step size}

\item{iter_warmup}{\code{\link[=integer]{integer()}} number of warmup iterations}

\item{iter_sampling}{\code{\link[=integer]{integer()}} number of sampling iterations}

\item{use_reduce_sum}{\code{\link[=logical]{logical()}} use within chain parallel threading}

\item{threads_per_chain}{\code{\link[=integer]{integer()}} number of threads per chain}

\item{...}{additional arguments to pass to \code{$sample()} method}
}
\value{
\code{\link[=fit]{fit()}}
}
\description{
Fit Markov Movement Model via CmdStanR
}
\details{
TBD
}
