% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gamma_parameters}
\alias{gamma_parameters}
\title{Gamma Distribution Parameters}
\usage{
gamma_parameters(mu = NULL, sd = NULL, alpha = NULL, beta = NULL)
}
\arguments{
\item{mu}{\code{\link[=numeric]{numeric()}} mean values}

\item{sd}{\code{\link[=numeric]{numeric()}} standard deviations}

\item{alpha}{\code{\link[=numeric]{numeric()}} alpha parameters}

\item{beta}{\code{\link[=numeric]{numeric()}} beta parameters}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Compute gamma distribution mean, standard deviation, alpha and beta
parameters from (array) mean and standard deviation or alpha and beta
parameters.
}
\examples{
# Scalar mu
mu <- 0.05
l <- gamma_parameters(mu, 0.01)
hist(rgamma(10000, l$alpha, l$beta), breaks = 100)

# Array mu
mu <- array(0.05, dim = c(2, 10, 3))
l <- gamma_parameters(mu, 0.001)
hist(rgamma(10000, l$alpha[1,1,1], l$beta[1,1,1]), breaks = 100)

}
