% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_n_to_t}
\alias{index_n_to_t}
\title{Index N to T}
\usage{
index_n_to_t(year_start, year_end, step_interval)
}
\arguments{
\item{year_start}{\code{\link[=integer]{integer()}} year of first tag release}

\item{year_end}{\code{\link[=integer]{integer()}} year of final tag recovery}

\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\link[=vector]{integer()}
}
\description{
Index N to T
}
\examples{

index_n_to_t(2011, 2018, "month")
index_n_to_t(2011, 2018, "quarter")
index_n_to_t(2011, 2018, "year")

}
