% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_t_to_year}
\alias{index_t_to_year}
\title{Index T to Years}
\usage{
index_t_to_year(year_start, year_end, step_interval)
}
\arguments{
\item{year_start}{\code{\link[=integer]{integer()}} year of first tag release}

\item{year_end}{\code{\link[=integer]{integer()}} year of final tag recovery}

\item{step_interval}{\code{\link[=character]{character()}} one of \code{"month"}, \code{"quarter"}
or \code{"year"}}
}
\value{
\link[=vector]{integer()}
}
\description{
Index T to Years
}
\examples{

index_t_to_year(2011, 2018, "month")
index_t_to_year(2011, 2018, "quarter")
index_t_to_year(2011, 2018, "year")

}
