% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{movement_index}
\alias{movement_index}
\title{Movement Index Matrix}
\usage{
movement_index(n, pattern = NULL, allow = NULL, disallow = NULL)
}
\arguments{
\item{n}{\code{\link[=integer]{integer()}} Number of areas.}

\item{pattern}{\code{\link[=integer]{integer()}} One of \code{1}: full movement, or \code{0}:
direct movement between numerically sequential neighbours only.}

\item{allow}{\code{\link[=integer]{integer()}} \code{\link[=matrix]{matrix()}} Each row indicates directional
movement between a pair of areas indexed from one.}

\item{disallow}{\code{\link[=integer]{integer()}} \code{\link[=matrix]{matrix()}} As for \code{allow}, but
specified movement is disallowed.}
}
\value{
A square binary matrix
}
\description{
Movement Index Matrix
}
\examples{

# Neighbours (default)
movement_index(6)

# Full movement
movement_index(6, 1)

# Neighbours plus 1-6, 2-5, and 5-2, but not 6-1
allow <- matrix(c(1,6,2,5,5,2), ncol = 2, byrow = TRUE)
movement_index(6, 0, allow)

# Full minus 1-6
disallow <- matrix(c(1,6), ncol = 2, byrow = TRUE)
movement_index(6, 1, disallow = disallow)

}
