% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.mmmfit}
\alias{plot.mmmfit}
\title{Plot Posterior Summaries}
\usage{
\method{plot}{mmmfit}(
  x = NULL,
  pars = c("p", "h", "phi"),
  ci = c(0.8, 0.95),
  area_names = NULL,
  group_names = NULL,
  ...
)
}
\arguments{
\item{x}{\code{\link[=mmmfit]{mmmfit()}} model fit}

\item{pars}{\code{\link[=character]{character()}} \code{\link[=vector]{vector()}}}

\item{ci}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}}}

\item{area_names}{\code{\link[=character]{character()}} \code{\link[=vector]{vector()}}}

\item{group_names}{\code{\link[=character]{character()}} \code{\link[=vector]{vector()}}}

\item{...}{not currently implemented}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} \code{\link[ggpubr:ggarrange]{ggpubr::ggarrange()}}
}
\description{
Plot Posterior Summaries
}
