% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_from_path}
\alias{read_from_path}
\title{Read File From Path}
\usage{
read_from_path(path)
}
\arguments{
\item{path}{\code{\link[=character]{character()}} file path}
}
\value{
the object at the file path
}
\description{
Read File From Path
}
