% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{strings_to_indices}
\alias{strings_to_indices}
\title{Extract Indices From A Character String}
\usage{
strings_to_indices(x, index = NULL)
}
\arguments{
\item{x}{\code{\link[=character]{character()}}}

\item{index}{\code{\link[=numeric]{numeric()}}}
}
\value{
\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}}
}
\description{
Extract Indices From A Character String
}
\examples{
x <- c("h[1,1,1]", "h[1,1,2]")
strings_to_indices(x)
strings_to_indices(x, index = 3)

x <- list("h[1,1,1]", "h[1,1,2]")
strings_to_indices(x)
strings_to_indices(x, index = 3)

}
