% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_priors}
\alias{summarise_priors}
\title{Summarise Priors}
\usage{
summarise_priors(s, d)
}
\arguments{
\item{s}{\code{\link[=list]{list()}} posterior draw summaries}

\item{d}{\code{\link[=list]{list()}} data}
}
\value{
\code{\link[=list]{list()}} prior summaries
}
\description{
Summarise Priors
}
