% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summary_to_tibbles}
\alias{summary_to_tibbles}
\title{Convert Movement Model Summary To List Of Tibbles}
\usage{
summary_to_tibbles(x)
}
\arguments{
\item{x}{\code{\link[cmdstanr:fit-method-summary]{cmdstanr::summary()}} from \code{\link[=mmmfit]{mmmfit()}} object}
}
\value{
\code{\link[=list]{list()}} of \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Convert Movement Model Summary To List Of Tibbles
}
