% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tag_arrays}
\alias{tag_arrays}
\title{Tag Arrays}
\usage{
tag_arrays(
  released,
  recovered,
  released_time_unit,
  released_time_max,
  liberty_time_max,
  liberty_days_min,
  colname_released_date,
  colname_released_area,
  colname_group,
  colname_recovered_date,
  colname_recovered_area,
  colname_id,
  area_list,
  group_list,
  released_date_start,
  released_date_end
)
}
\arguments{
\item{released}{\code{\link[=data.frame]{data.frame()}} released tags. See details.}

\item{recovered}{\code{\link[=data.frame]{data.frame()}} recovered tags. See details.}

\item{released_time_unit}{\code{\link[=character]{character()}} one of \code{year}, \code{quarter},
or \code{month}.}

\item{released_time_max}{\code{\link[=numeric]{numeric()}} equivalent to \code{T_released}.}

\item{liberty_time_max}{\code{\link[=numeric]{numeric()}} equivalent to \code{T_liberty}.}

\item{liberty_days_min}{\code{\link[=numeric]{numeric()}} (not currently implemented).}

\item{colname_released_date}{\code{\link[=character]{character()}} released date column name.}

\item{colname_released_area}{\code{\link[=character]{character()}} released area column name.}

\item{colname_group}{\code{\link[=character]{character()}} released group column name.}

\item{colname_recovered_date}{\code{\link[=character]{character()}} recovered date column name.}

\item{colname_recovered_area}{\code{\link[=character]{character()}} recovered area column name.}

\item{colname_id}{\code{\link[=character]{character()}} tag identifier column name.}

\item{area_list}{\code{\link[=list]{list()}} of named atomic vectors. See details.}

\item{group_list}{\code{\link[=list]{list()}} of named atomic vectors. See details.}

\item{released_date_start}{\code{\link[=character]{character()}} earliest released date.}

\item{released_date_end}{\code{\link[=character]{character()}} latest released date.}
}
\value{
\code{\link[=list]{list()}} of released and recovered \code{\link[=array]{array()}}s
}
\description{
Tag Arrays
}
\details{
TBD
}
