% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_steps_per_term}
\alias{compute_steps_per_term}
\title{Compute The Number of Model Steps Per Term}
\usage{
compute_steps_per_term(x, y)
}
\arguments{
\item{x}{\code{\link[=character]{character()}} term interval. One of \code{"year"}, \code{"quarter"}
or \code{"month"}}
}
\value{
\code{\link[=integer]{integer()}}
}
\description{
Compute The Number of Model Steps Per Term
}
\examples{

compute_steps_per_term("month", "month")
compute_steps_per_term("month", "quarter")
compute_steps_per_term("month", "year")
compute_steps_per_term("quarter", "month")
compute_steps_per_term("quarter", "quarter")
compute_steps_per_term("quarter", "year")
compute_steps_per_term("year", "month")
compute_steps_per_term("year", "quarter")
compute_steps_per_term("year", "year")

}
