% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_sample_summary}
\alias{create_sample_summary}
\title{Create MCMC Sample Summary}
\usage{
create_sample_summary(x, ci = c(0.8, 0.95))
}
\arguments{
\item{x}{\code{\link[tibble:tibble]{tibble::tibble()}} sample tibble (see \code{\link[=create_sample_tibble]{create_sample_tibble()}})}

\item{ci}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of credible interval levels}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create MCMC Sample Summary
}
