% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_tag_array}
\alias{create_tag_array}
\title{Create Tag Array}
\usage{
create_tag_array(
  tag_data,
  list_regions,
  list_sizes,
  year_released_start,
  year_recovered_end,
  step_interval = "month",
  step_duration_max = NULL,
  days_duration_min = 30,
  colname_date_released = "date_released",
  colname_date_recovered = "date_recovered",
  colname_region_released = "region_released",
  colname_region_recovered = "region_recovered",
  colname_size_released = "size_released"
)
}
\arguments{
\item{tag_data}{\code{\link[=data.frame]{data.frame()}}}

\item{list_regions}{\code{\link[=list]{list()}}}

\item{list_sizes}{\code{\link[=list]{list()}}}

\item{year_released_start}{\code{\link[=integer]{integer()}} year that the first tag was released}

\item{year_recovered_end}{\code{\link[=integer]{integer()}} year the last tag was recovered}

\item{step_interval}{\code{\link[=character]{character()}} one of \code{"year"}, \code{"quarter"}
or \code{"month"}}

\item{step_duration_max}{\code{\link[=integer]{integer()}} optional constraint on the maximum
steps duration at large}

\item{days_duration_min}{\code{\link[=integer]{integer()}} minimum number of days before recovery}

\item{colname_date_released}{\code{\link[=character]{character()}}}

\item{colname_date_recovered}{\code{\link[=character]{character()}}}

\item{colname_region_released}{\code{\link[=character]{character()}}}

\item{colname_region_recovered}{\code{\link[=character]{character()}}}

\item{colname_size_released}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[=array]{array()}}
}
\description{
Create Tag Array
}
