% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit}
\alias{fit}
\title{Fit A Markov Movement Model via CmdStanR}
\usage{
fit(
  tag_data,
  list_regions,
  list_sizes,
  year_released_start,
  year_recovered_end,
  step_liberty_max = NULL,
  term_interval = "quarter",
  colname_date_released = "date_released",
  colname_date_recovered = "date_recovered",
  colname_region_released = "region_released",
  colname_region_recovered = "region_recovered",
  colname_size_released = "size_released",
  movement_pattern = 2,
  movement_allow = NULL,
  movement_disallow = NULL,
  mu_mortality_rate = rep(0.1, length(list_regions)),
  mu_reporting_rate = rep(1, length(list_regions)),
  mu_fishing_mean_rate = rep(0.1, length(list_regions)),
  mu_initial_loss_rate = 0.1,
  mu_ongoing_loss_rate = 0.02,
  mu_dispersion = 1,
  mu_fishing_time_deviation = NULL,
  sd_mortality_rate = 0.1 * mu_mortality_rate,
  sd_reporting_rate = 0.1 * mu_reporting_rate,
  sd_fishing_mean_rate = 0.1 * mu_fishing_mean_rate,
  sd_initial_loss_rate = 0.1 * mu_initial_loss_rate,
  sd_ongoing_loss_rate = 0.1 * mu_ongoing_loss_rate,
  sd_dispersion = 0.25 * mu_dispersion,
  cv_movement_time_deviation = 0.1,
  cv_movement_term_deviation = 0.1,
  cv_movement_size_deviation = 0.1,
  cv_fishing_time_deviation = 0.1,
  cv_fishing_term_deviation = 0.1,
  cv_fishing_size_deviation = 0.1,
  expected_fudge = 1e-12,
  data = NULL,
  chains = 1,
  step_size = 0.01,
  iter_warmup = 250,
  iter_sampling = 750,
  use_reduce_sum = FALSE,
  threads_per_chain = 8,
  ...
)
}
\arguments{
\item{tag_data}{\code{\link[=data.frame]{data.frame()}}}

\item{list_regions}{\code{\link[=list]{list()}}}

\item{list_sizes}{\code{\link[=list]{list()}}}

\item{year_released_start}{\code{\link[=integer]{integer()}} year that the first tag was released}

\item{year_recovered_end}{\code{\link[=integer]{integer()}} year that the last tag was recovered}

\item{step_liberty_max}{\code{\link[=integer]{integer()}}}

\item{term_interval}{\code{\link[=character]{character()}} in \code{c("year", "month", "quarter")}}

\item{colname_date_released}{\code{\link[=character]{character()}}}

\item{colname_date_recovered}{\code{\link[=character]{character()}}}

\item{colname_region_released}{\code{\link[=character]{character()}}}

\item{colname_region_recovered}{\code{\link[=character]{character()}}}

\item{colname_size_released}{\code{\link[=character]{character()}}}

\item{movement_pattern}{\code{\link[=integer]{integer()}} One of \code{1}: movement between
all pairs of regions at each step, or \code{2}:
movement between numerically sequential regions (the default)}

\item{movement_allow}{\code{\link[=integer]{integer()}} \code{\link[=matrix]{matrix()}} Adjusts the movement pattern
to allow additional movement between regions. Each row indicates
directional movement between a pair of regions}

\item{movement_disallow}{\code{\link[=integer]{integer()}} \code{\link[=matrix]{matrix()}} As for
\code{movement_allow}, but specified movement is disallowed}

\item{mu_mortality_rate}{\code{\link[=numeric]{numeric()}}}

\item{mu_reporting_rate}{\code{\link[=numeric]{numeric()}}}

\item{mu_fishing_mean_rate}{\code{\link[=numeric]{numeric()}}}

\item{mu_initial_loss_rate}{\code{\link[=numeric]{numeric()}}}

\item{mu_ongoing_loss_rate}{\code{\link[=numeric]{numeric()}}}

\item{mu_dispersion}{\code{\link[=numeric]{numeric()}}}

\item{mu_fishing_time_deviation}{\code{\link[=numeric]{numeric()}}}

\item{sd_mortality_rate}{\code{\link[=numeric]{numeric()}}}

\item{sd_reporting_rate}{\code{\link[=numeric]{numeric()}}}

\item{sd_fishing_mean_rate}{\code{\link[=numeric]{numeric()}}}

\item{sd_initial_loss_rate}{\code{\link[=numeric]{numeric()}}}

\item{sd_ongoing_loss_rate}{\code{\link[=numeric]{numeric()}}}

\item{sd_dispersion}{\code{\link[=numeric]{numeric()}}}

\item{cv_movement_time_deviation}{\code{\link[=numeric]{numeric()}}}

\item{cv_movement_term_deviation}{\code{\link[=numeric]{numeric()}}}

\item{cv_movement_size_deviation}{\code{\link[=numeric]{numeric()}}}

\item{cv_fishing_time_deviation}{\code{\link[=numeric]{numeric()}}}

\item{cv_fishing_term_deviation}{\code{\link[=numeric]{numeric()}}}

\item{cv_fishing_size_deviation}{\code{\link[=numeric]{numeric()}}}

\item{expected_fudge}{\code{\link[=numeric]{numeric()}}}

\item{data}{\code{\link[=list]{list()}} See details}

\item{chains}{\code{\link[=integer]{integer()}} number of chains}

\item{step_size}{\code{\link[=numeric]{numeric()}} initial step size}

\item{iter_warmup}{\code{\link[=integer]{integer()}} number of warmup iterations}

\item{iter_sampling}{\code{\link[=integer]{integer()}} number of sampling iterations}

\item{use_reduce_sum}{\code{\link[=logical]{logical()}} use within chain parallel threading}

\item{threads_per_chain}{\code{\link[=integer]{integer()}} number of threads per chain}

\item{...}{additional arguments to pass to \code{$sample()} method}
}
\value{
\code{\link[=fit]{fit()}}
}
\description{
Stan model data may be specified in two ways: via arguments \code{tag_data}
through \code{expected_fudge} several of which take advantage of helper
functions and many of which have sensible defaults, xor via the \code{data}
argument which takes precedent when present and must be compatible with the
Stan model directly.
}
\details{
The \code{data} argument can offer greater flexibility, for example
the times and terms that define the step size used in the model
are currently constrained to times = years and terms in
\code{c("year", "month", "quarter")} by the helper functions. By specifying
\code{tags_released} and \code{tags_recovered} manually in the \code{data}
argument and adjusting the corresponding index limits, arbitrary time and
step units are possible.

TBD
}
