% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmfit}
\alias{mmmfit}
\title{Fit Markov Movement Model via RStan}
\usage{
mmmfit(data, iter = 750, chains = 1, warmup = 250)
}
\arguments{
\item{data}{\code{\link[=list]{list()}}}
}
\value{
\code{\link[=mmmfit]{mmmfit()}}
}
\description{
Fit Markov Movement Model via RStan
}
