% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmfit}
\alias{mmmfit}
\title{Fit Markov Movement Model via RStan}
\usage{
mmmfit(
  data,
  chains = 1,
  iter_warmup = 250,
  iter_sampling = 750,
  use_reduce_sum = FALSE,
  threads_per_chain = 8,
  ...
)
}
\arguments{
\item{data}{\code{\link[=list]{list()}}}

\item{chains}{\code{\link[=integer]{integer()}} number of chains}

\item{iter_warmup}{\code{\link[=integer]{integer()}} number of warmup iterations}

\item{iter_sampling}{\code{\link[=integer]{integer()}} number of sampling iterations}

\item{use_reduce_sum}{\code{\link[=logical]{logical()}} use within chain parallel threading}

\item{threads_per_chain}{\code{\link[=integer]{integer()}} number of threads per chain}

\item{...}{additional arguments to pass to \code{$sample()} method}
}
\value{
\code{\link[=mmmfit]{mmmfit()}}
}
\description{
Fit Markov Movement Model via RStan
}
